tests/wpt/web-platform-tests/html/user-activation/activation-hierarchy-sameorigin-parent-manual.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
  <h1>User activation visibility across same-origin frame boundary</h1>
  <h2>Confirms that the state is visible in both parent and child</h2>
  <ol id="instructions">
    <li>Click anywhere on the green area (child frame).
  </ol>
  <iframe id="child" width="400" height="400"></iframe>
  <script>
    async_test(function(t) {
        assert_false(navigator.userActivation.isActive);
        assert_false(navigator.userActivation.hasBeenActive);

        let child = document.getElementById("child");

        window.addEventListener("message", t.step_func(event => {
            if (event.source === frames[0].frames[0] && event.data === 'checked') {
                // Parent should be active after child is clicked.
                assert_true(navigator.userActivation.isActive);
                assert_true(navigator.userActivation.hasBeenActive);
                t.done();
            }
        }));
        child.src = "resources/activation-hierarchy-sameorigin-child.html";
    }, "Parent test");
  </script>
</body>
</html>
